# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/mindmaps/generate": {
      "post": {
        "tags": [
          "mindmaps"
        ],
        "summary": "Generate Mindmap",
        "operationId": "generate_mindmap_mindmaps_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MindMapGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MindMapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/mindmaps/{mindmap_id}": {
      "get": {
        "tags": [
          "mindmaps"
        ],
        "summary": "Get Mindmap",
        "operationId": "get_mindmap_mindmaps__mindmap_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mindmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mindmap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MindMapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "mindmaps"
        ],
        "summary": "Update Mindmap",
        "operationId": "update_mindmap_mindmaps__mindmap_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mindmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mindmap Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MindMapUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MindMapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mindmaps/{mindmap_id}/export": {
      "get": {
        "tags": [
          "mindmaps"
        ],
        "summary": "Export Mindmap",
        "operationId": "export_mindmap_mindmaps__mindmap_id__export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mindmap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mindmap Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|mermaid)$",
              "default": "json",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MindMapEdge": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "target"
        ],
        "title": "MindMapEdge"
      },
      "MindMapGenerateRequest": {
        "properties": {
          "source_type": {
            "type": "string",
            "enum": [
              "note",
              "knowledge",
              "mixed"
            ],
            "title": "Source Type"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "max_nodes": {
            "type": "integer",
            "title": "Max Nodes",
            "default": 40
          },
          "max_depth": {
            "type": "integer",
            "title": "Max Depth",
            "default": 4
          },
          "focus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focus"
          }
        },
        "type": "object",
        "required": [
          "source_type",
          "source_ids"
        ],
        "title": "MindMapGenerateRequest"
      },
      "MindMapNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "level": {
            "type": "integer",
            "title": "Level",
            "default": 0
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "concept"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "source_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Refs",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "MindMapNode"
      },
      "MindMapResponse": {
        "properties": {
          "mindmap_id": {
            "type": "string",
            "title": "Mindmap Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/MindMapNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/MindMapEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/SourceCitation"
            },
            "type": "array",
            "title": "Citations",
            "default": []
          },
          "source_refs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Source Refs",
            "default": []
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "mindmap_id",
          "title",
          "source_type",
          "source_ids",
          "nodes",
          "edges"
        ],
        "title": "MindMapResponse"
      },
      "MindMapUpdateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/MindMapNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/MindMapEdge"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "required": [
          "title",
          "nodes",
          "edges"
        ],
        "title": "MindMapUpdateRequest"
      },
...
      "SourceCitation": {
        "properties": {
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id"
          },
          "quote": {
            "type": "string",
            "title": "Quote"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "source_type",
          "source_id",
          "title",
          "quote"
        ],
        "title": "SourceCitation"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}